refactored components location and objects definition
