+++ src/ui/Detail/Detail.tsx
@@ -1,4 +1,6 @@
-import { Text, View, Image, Button } from 'react-native'
+import { Text, Image, Button } from 'react-native'
+
+import { Box } from '../_components/_objects/Box'
 
 type Props = {
   title: string
@@ -8,10 +10,10 @@ type Props = {
 }
 
 export const Detail = ({ title, description, image, onPress }: Props) => (
-  <View testID="detailScreen">
+  <Box vertical="spaceAround" testID="detailScreen">
     <Image testID="articleImage" style={{ height: 50, width: 50 }} source={{ uri: image }} />
     <Text>{title}</Text>
     <Text>{description}</Text>
     <Button testID="getArticleButton" title="Open in web" onPress={onPress} />
-  </View>
+  </Box>
 )
